Refactor rllab-ros
- Add more introduction for the project.
- In current design, we separate every spesific task/experiment env, and give them separate detailed implementation.
- The RosEnv would have more functionality to control gazebo for simulation in the future.
