+++ contrib/ros/envs/sawyer/task_object_manager.py
